refactor(procedures): Move Procedure container and App to distinct files
