+++ packages/cozy-procedures/src/Procedure.jsx
@@ -1,16 +1,8 @@
 import React from 'react'
-import { Provider, connect } from 'react-redux'
 import get from 'lodash/get'
-
-import store from './redux/store'
-import Context from './redux/context'
-import creditApplicationTemplate from './templates/creditApplicationTemplate'
-import {
-  init as initPersonalData,
-  fetchMyself
-} from './redux/personalDataSlice'
-import withLocales from './withLocales'
 import { withClient } from 'cozy-client'
+
+import creditApplicationTemplate from './templates/creditApplicationTemplate'
 
 class Procedure extends React.Component {
   componentDidMount() {
@@ -25,24 +17,4 @@ class Procedure extends React.Component {
   }
 }
 
-const mapDispatchToProps = {
-  initPersonalData,
-  fetchMyself
-}
-
-const ProcedureContainer = withLocales(
-  connect(
-    null,
-    mapDispatchToProps,
-    null,
-    { context: Context }
-  )(Procedure)
-)
-
-const ProcedureWithStore = props => (
-  <Provider store={store} context={Context}>
-    <ProcedureContainer {...props} />
-  </Provider>
-)
-
-export default withClient(ProcedureWithStore)
+export default withClient(Procedure)
