+++ packages/cozy-procedures/src/index.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import { Route, IndexRoute } from 'react-router'
 
-import Procedure from './Procedure'
+import App from './App'
 import Overview from './containers/Overview'
 import Amount from './components/Amount'
 import Duration from './components/Duration'
@@ -9,7 +9,7 @@ import PersonalDataForm from './containers/PersonalDataForm'
 import Documents from './components/Documents'
 
 const injectProcedureRoutes = ({ root }) => (
-  <Route path={root} component={Procedure}>
+  <Route path={root} component={App}>
     <IndexRoute component={() => <Overview />} />
     <Route path="amount" component={() => <Amount />} />
     <Route path="duration" component={() => <Duration />} />
